python-service/requirements.txt
fastapi
uvicorn
pypdf
python-multipart
google-generativeai
python-dotenv
python-docx
